requirements.txt
breadability==0.1.20
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.0.11
docopt==0.6.2
idna==2.6
imageio==2.1.2
lxml==4.6.2
moviepy==0.2.3.2
nltk==3.5
numpy==1.20.1
olefile==0.46
Pillow==8.1.0
pysrt==1.1.2
pytube==10.4.1
pandas==1.2.2
plotly-express==0.4.1
requests==2.25.1
six==1.15.0
sumy==0.8.1
tqdm==4.11.2
urllib3==1.26.3
youtube-dl==2021.2.10
streamlit==0.76.0